website/requirements.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
pika==1.2.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
